requirements.txt
Flask
flask-cors
numpy
scipy
PyWavelets
opencv-python
Pillow
scikit-image